move seed user and workspace to manage.ts seed action
